rename sub example snippets
